+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display `<redtext` in posts.
 - Added CSS variables for greentext and redtext colors.
+- Added pagination to gallery.
 
 ### Changed